added tournament id to url in registration form
