+++ pages/api/user/create.ts
@@ -7,18 +7,16 @@ export default async function create(
 ) {
   if (req.method === "POST") {
     const playerData = JSON.parse(req.body);
-    const t = await prisma.tournament.findFirst({});
-
     const result = await prisma.user.create({
       data: {
         firstName: playerData.firstName,
         lastName: playerData.lastName,
         phone: playerData.phone,
         email: playerData.email,
-        tournament: { connect: { id: t.id } },
+        tournament: { connect: { id: playerData.tournamentId } },
         player: {
           create: {
-            tournament: { connect: { id: t.id } },
+            tournament: { connect: { id: playerData.tournamentId } },
             alias: playerData.alias,
             address: playerData.address,
             learningInstitution: playerData.learningInstitution,
